Add support to zero metric on not matching message
If a metric never gets set sets, it won't be zero but simply
non-existing. A error counter for instance is either non-existing or >=1
which means we won't be able to alert on it increasing from 0.

This change makes it possible to set all metrics to 0 for which we
receive a message which doesn't match the matcher.